gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
azure-ai-vision-imageanalysis==1.0.0
azure-ai-documentintelligence==1.0.0
azure-core==1.30.2
pydantic==2.7.4
